feat: Simple load balanced rate control
By introducing the option `balanced: true` we interleave the initial ring w/ the interval divided by size so all calls will have some air between each call e.g. a simple load balancing.

---

By default all calls to the `limit_method` will be bursted, e.g. as quick as possible, until the rate is exceeded.
Then we wait for the remainder of the interval to continue. To even out the burst, an optional `balanced` parameter can be
provided to enable interleaving between the method calls, e.g: `interleave = interval / size`.

``` ruby
  ...
  limit_method :tick, rate: 60, balanced: true
  ...
```

For example: with an interval of 60 seconds and a rate of 60:

`balanced: false` As quickly as possible we call the method 60 times, then we wait for the remainder of the time.

`balanced: true` We interleave each call with 1 second so we call this method every second.

---

Considerations before merging this PR:

- Use `interleaved` instead of `balanced` as an option name
- Make the interleave configurable: e.g. instead of `true/false` provide
  an `float/integer` (which also favours renaming to `interleave`)
- ditch this PR because when this is an issue your rate/interval is wrong.